refactor: moved Image to components directory
