+++ src/App.js
@@ -1,7 +1,6 @@
 import logo from './logo.svg';
 import './App.css';
-
-import Image from './Image';
+import { Image } from './components/image';
 
 const src = 'https://i.redd.it/2ghbw4ea28471.jpg';
 
@@ -10,7 +9,7 @@ function App() {
     <div className="App">
       <img src={logo} className="App-logo" alt="logo" />
 
-      <Image src={src} alt="image" height="200px" placeholderSrc={logo} />
+      <Image src={src} alt="image" height="200" />
     </div>
   );
 }
