+++ src/components/image/index.js
@@ -1,0 +1,2 @@
+export * from './image';
+export * from './use-image';
